requirements.txt
amqp==5.1.1
asgiref==3.6.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
boto==2.49.0
boto3==1.26.116
botocore==1.29.116
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
Django==4.2
django-amazon-ses==4.0.1
django-analytical==3.1.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-matplotlib==0.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-storages==1.13.2
django-taggit==4.0.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.5
exponent-server-sdk==2.0.0
fcm-django==2.0.0
firebase-admin==6.1.0
fonttools==4.39.4
geographiclib==2.0
geopy==2.3.0
gevent==22.10.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
idna==3.4
importlib-resources==6.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlip==0.2
mixpanel==4.10.0
mixpanel-python==0.0.2
msgpack==1.0.5
nltk==3.8.1
numpy==1.24.3
openapi-codec==1.3.2
packaging==23.1
pandas==2.0.3
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
vine==5.0.0
wcwidth==0.2.6
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
